feat(dispatch): close PROB-050 A-4+A-5+A-6+A-7+A-11+A-15 — claude_print refactor
PROB-050 follow-up sprint per ADR-011 §Phase B. 6 sub-items closed in one
focused refactor of the dispatch layer; the duplication between
AgentDispatcher and PluginDispatcher (identical 9-step claude --print
recipe per ADR-011 §Decision) collapses into a single source of truth.

Context: PR-B (#243, merged) closed A-14 (CWE-426 binary substitution).
This PR closes the next 6 architectural items from the original PROB-050
acceptance criteria. The remaining items (A-1 spec bumps, A-2 ADR-010
amendment, A-3 #[ignore] e2e, A-8/A-9/A-10/A-12/A-13/A-16..A-29) are
either methodology hygiene or larger refactors and stay open under
PROB-050 / PROB-051 follow-ups for v0.30.0.

A-4: claude_print::invoke()
  Extracted the 9-step recipe (argv build + env allowlist + prompt-via-stdin
  + spawn + timeout + parse + render) into a single `pub(super) async fn
  invoke(label, slug, step, workspace, binary, timeout)`. Both dispatchers
  reduce to (a) variant unpack, (b) compute slug + label, (c) resolve
  binary, (d) call invoke. Unifies the parse-failure diagnostic to combine
  exit_code (agent's pattern) + stdout_preview (plugin's pattern) — union
  of pre-A-4 dispatcher diagnostics. Plugin's "failed to decode JSON
  envelope" wording preserved so existing test assertion still passes.

  Net effect: agent_dispatcher.rs body shrank from 442 to 210 lines
  (-232 LOC); plugin_dispatcher.rs from 348 to 153 (-195). The 426-line
  reduction in dispatcher bodies is partially offset by 326 added lines
  in claude_print.rs (the new invoke + helpers); net delta is +46 LOC
  with massively reduced duplication.

A-5: which_in_path consolidation
  Promoted `helpers::which_in_path` from `fn` to `pub(super) fn`. Removed
  3 identical local copies (helpers.rs:303, agent_dispatcher.rs:369,
  plugin_dispatcher.rs:323). All call sites now consume the single source
  via `super::helpers::which_in_path` (or `super::super::helpers::...`
  from inside test modules).

A-6: shared DISPATCH_ENV_LOCK
  Moved `static ENV_GUARD: tokio::sync::Mutex<()>` from
  `agent_dispatcher::tests` (line ~396) to a shared
  `pub(super) static DISPATCH_ENV_LOCK` in `claude_print.rs` (under
  `#[cfg(test)]` so it doesn't ship in release binaries). Now consumed by
  `agent_dispatcher::tests` AND `helpers::tests` (the latter previously
  had NO env-mutation lock — Round 5 audit Logic LOW-1 race risk closed).
  `helpers::tests::resolve_forgeplan_binary_respects_env_override`
  converted to `#[tokio::test]` to await the shared async lock.

A-7: visibility tightening
  Empirically verified (`rg <name> crates/`) that no external library
  consumer reads claude_print's pub items. Tightened:
    - DEFAULT_BUDGET_USD, DEFAULT_ALLOWED_TOOLS: pub → pub(crate)
    - ClaudePrintResponse: pub → pub(super)
    - assemble_prompt, add_dir_for_produces_at: pub → pub(super)
    - effective_allowed_tools, effective_budget_usd: pub → pub(super)
  Added #[allow(dead_code)] on ClaudePrintResponse.duration_ms +
  session_id (preserved for future telemetry; clippy -D dead_code would
  otherwise block the lockdown).

A-11: parse_envelope + format_timeout_msg
  Extracted both into claude_print.rs. parse_envelope uses the trimmed
  pattern (was: agent .trim(), plugin no-trim → Round 5 LOW-2 silent drift
  risk). format_timeout_msg uses .as_secs() on both paths (was: agent
  Debug repr, plugin .as_secs()). Single source of truth.

A-15: build_argv
  Extracted `pub(super) fn build_argv(slug, step, workspace_root)
  -> Result<Vec<String>, String>` with both security gates inline
  (validate_allowed_tools + add_dir_for_produces_at). Argv-shape parity
  between Plugin and Agent now enforced by construction, not convention.

Quality gates:
  cargo fmt --check                                          → 0 diff
  cargo clippy --workspace --all-targets \
     --features test-helpers -- -D warnings                  → 0 warnings
  cargo test --workspace --features test-helpers             → 1974 passed,
                                                               0 failed
                                                               (unchanged)
  scripts/check-mcp-tool-count.sh                            → 0 drift
                                                               (still 63 tools)

Behavior preserved:
  - Argv shape byte-identical to pre-refactor (verified by existing
    fake-claude-binary tests in both dispatchers).
  - Error messages preserved where tests asserted on them (plugin's
    "failed to decode" still fires; agent's parse-failure diagnostic
    expanded to include stdout_preview that plugin had).
  - Timeout messages now consistent (.as_secs() on both — fixed the
    Debug-repr-leak in agent path).

Refs: PROB-050 (A-4, A-5, A-6, A-7, A-11, A-15), Round 5 audit
Logic LOW-1, ADR-011
