+++ crates/forgeplan-core/src/playbook/dispatch/claude_print.rs
@@ -24,23 +24,58 @@
 //! - Budget cap is mandatory — every invocation passes `--max-budget-usd`.
 //! - Stdin prompt is mandatory.
 
+use std::collections::HashMap;
 use std::path::Path;
 use std::sync::OnceLock;
+use std::time::Duration;
 
 use regex::Regex;
 use serde::Deserialize;
 
 use crate::playbook::types::Step;
 
+use super::{DispatchError, DispatchOutcome};
+
+/// Cross-dispatcher serialization lock for tests that mutate process-global
+/// env vars (PATH, FORGEPLAN_CLAUDE_BIN, FORGEPLAN_BIN).
+///
+/// PROB-050 A-6 closure: previously each dispatcher (`agent_dispatcher::tests`,
+/// `plugin_dispatcher::tests`, `helpers::tests`) had its own ENV_GUARD or no
+/// guard at all. `cargo test` runs tests on multiple threads in the SAME
+/// process; without a shared lock, fake-claude scripts in concurrently-running
+/// tests can see the temporarily-broken PATH set by `*_when_tool_absent`
+/// cases and lose their ability to locate `/bin/sh` for shebang exec
+/// (~1 in 5 runs flake). Round-5 audit Logic LOW-1 noted this race risk
+/// for `helpers::tests::resolve_forgeplan_binary_respects_env_override`
+/// which had no guard at all.
+///
+/// Uses `tokio::sync::Mutex` because every consumer holds the guard across
+/// `await` points (clippy::await_holding_lock would fire on `std::sync::Mutex`
+/// in `#[tokio::test]` contexts).
+///
+/// Test-only — gated behind `#[cfg(test)]` so it doesn't ship in release
+/// binaries. Visible to sibling modules (`agent_dispatcher`, `plugin_dispatcher`,
+/// `helpers`) via `pub(super)`.
+#[cfg(test)]
+pub(super) static DISPATCH_ENV_LOCK: tokio::sync::Mutex<()> = tokio::sync::Mutex::const_new(());
+
 /// Default per-invocation budget when `Step.budget_usd` is `None`.
 /// Matches ADR-011 §Decision point 2 ("default $1.00, configurable per step").
-pub const DEFAULT_BUDGET_USD: f64 = 1.00;
+///
+/// PROB-050 A-7 closure: tightened from `pub` to `pub(crate)`. Empirically
+/// (`rg DEFAULT_BUDGET_USD crates/`) no external library consumer reads this
+/// constant; restricting to crate-internal protects the value from being
+/// pinned as a stability contract by downstream crates that we'd then have
+/// to honour across release boundaries. Widen back to `pub` when a
+/// marketplace plugin author needs it.
+pub(crate) const DEFAULT_BUDGET_USD: f64 = 1.00;
 
 /// Default tool allowlist for analytic agents when `Step.allowed_tools` is
 /// `None`. Least-privilege: read-only filesystem + grep + glob.
-/// Wave 1 may revise based on real-step requirements (e.g. add `Write` for
-/// produces_at flows).
-pub const DEFAULT_ALLOWED_TOOLS: &[&str] = &["Read", "Glob", "Grep"];
+///
+/// PROB-050 A-7 closure: tightened from `pub` to `pub(crate)` (same reason
+/// as `DEFAULT_BUDGET_USD`).
+pub(crate) const DEFAULT_ALLOWED_TOOLS: &[&str] = &["Read", "Glob", "Grep"];
 
 /// Structured envelope returned by `claude --print --output-format json`.
 /// EVID-093 documented 17 fields; this struct captures the load-bearing
@@ -51,7 +86,11 @@ pub const DEFAULT_ALLOWED_TOOLS: &[&str] = &["Read", "Glob", "Grep"];
 /// them.
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "snake_case")]
-pub struct ClaudePrintResponse {
+/// PROB-050 A-7 closure: tightened from `pub` to `pub(super)`.
+/// Used only inside the dispatch module (parse_envelope returns it,
+/// invoke consumes it). External library consumers would couple to
+/// claude CLI's private envelope shape — high churn risk.
+pub(super) struct ClaudePrintResponse {
     /// `true` iff `claude` itself reported an error condition (API error,
     /// budget exceeded mid-flight, internal failure). The exit code alone
     /// does not disambiguate.
@@ -69,10 +108,19 @@ pub struct ClaudePrintResponse {
     #[serde(default)]
     pub total_cost_usd: f64,
     /// Wall-clock duration in milliseconds.
+    ///
+    /// PROB-050 A-7 closure: marked `#[allow(dead_code)]` because tightening
+    /// the parent struct to `pub(super)` revealed there are no current
+    /// readers — the field is preserved for future telemetry / log enrichment
+    /// (Round 5 deferred to PROB-051 D-class). Without `#[allow]`, clippy
+    /// `-D dead_code` blocks the lockdown.
     #[serde(default)]
+    #[allow(dead_code)]
     pub duration_ms: u64,
     /// Session id for downstream correlation / debugging.
+    /// PROB-050 A-7: see `duration_ms` rationale.
     #[serde(default)]
+    #[allow(dead_code)]
     pub session_id: Option<String>,
 }
 
@@ -126,7 +174,7 @@ impl ClaudePrintResponse {
 /// 2. If `step.produces_at` is set, appends:
 ///    `Write output to \`<produces_at>\` using the Write tool.`
 /// 3. Returns the assembled string ready for stdin pipe.
-pub fn assemble_prompt(step: &Step) -> String {
+pub(super) fn assemble_prompt(step: &Step) -> String {
     let task = step
         .input
         .as_ref()
@@ -169,7 +217,7 @@ pub fn assemble_prompt(step: &Step) -> String {
 /// `Component::ParentDir` are rejected. Caller (PluginDispatcher /
 /// AgentDispatcher) maps the error to `DispatchError::Transport` and refuses
 /// to spawn `claude`.
-pub fn add_dir_for_produces_at(
+pub(super) fn add_dir_for_produces_at(
     step: &Step,
     workspace_root: &Path,
 ) -> Result<Option<std::path::PathBuf>, String> {
@@ -253,7 +301,7 @@ pub(crate) fn format_budget(budget: f64) -> String {
 /// Resolve the effective tool allowlist: `Step.allowed_tools` if `Some`,
 /// otherwise [`DEFAULT_ALLOWED_TOOLS`]. Returns owned `String`s so the
 /// caller can pass them to `Command::args` without lifetime gymnastics.
-pub fn effective_allowed_tools(step: &Step) -> Vec<String> {
+pub(super) fn effective_allowed_tools(step: &Step) -> Vec<String> {
     step.allowed_tools.clone().unwrap_or_else(|| {
         DEFAULT_ALLOWED_TOOLS
             .iter()
@@ -264,8 +312,268 @@ pub fn effective_allowed_tools(step: &Step) -> Vec<String> {
 
 /// Resolve the effective budget: `Step.budget_usd` if `Some`, otherwise
 /// [`DEFAULT_BUDGET_USD`].
-pub fn effective_budget_usd(step: &Step) -> f64 {
+pub(super) fn effective_budget_usd(step: &Step) -> f64 {
     step.budget_usd.unwrap_or(DEFAULT_BUDGET_USD)
+}
+
+/// Parse the JSON envelope returned by `claude --print --output-format json`.
+///
+/// PROB-050 A-11 closure: pre-fix PluginDispatcher used
+/// `serde_json::from_str(&stdout_text)` (no trim), AgentDispatcher used
+/// `serde_json::from_str(stdout_str.trim())`. The trimmed variant tolerates
+/// trailing newlines + BOM that real `claude --print` can emit, so it's
+/// the safer pattern. Both dispatchers now consume this helper.
+///
+/// # Errors
+///
+/// Returns `serde_json::Error` when stdout is not parseable as a
+/// [`ClaudePrintResponse`]. Callers wrap in dispatcher-specific
+/// diagnostics (with agent name / plugin/target context).
+pub(super) fn parse_envelope(stdout: &[u8]) -> Result<ClaudePrintResponse, serde_json::Error> {
+    let s = String::from_utf8_lossy(stdout);
+    serde_json::from_str(s.trim())
+}
+
+/// Format the dispatcher timeout-error message in seconds.
+///
+/// PROB-050 A-11 closure: pre-fix PluginDispatcher used `.as_secs()` →
+/// `"plugin `foo/bar` timed out after 300s"`. AgentDispatcher used `{:?}`
+/// (Debug repr) → `"agent `foo` timed out after 300s"` for whole seconds
+/// but `"300.500s"` for fractional — leaks Duration's internal layout into
+/// user-visible diagnostics. Single source of truth on `.as_secs()`
+/// format; both dispatchers consume this helper.
+///
+/// `label` is the dispatcher-specific prefix (e.g. `agent \`foo\``,
+/// `plugin \`foo/bar\``). Helper returns the full sentence.
+pub(super) fn format_timeout_msg(label: &str, duration: std::time::Duration) -> String {
+    format!("{label} timed out after {}s", duration.as_secs())
+}
+
+/// Build the full argv for `claude --print` invocation, with all security
+/// validation gates applied.
+///
+/// PROB-050 A-15 closure: extracted from AgentDispatcher::dispatch and
+/// PluginDispatcher::dispatch (identical 11-step recipes per ADR-011
+/// §Decision). Single source of truth — argv-shape tests live alongside
+/// this function in `claude_print::tests` (no fake-binary indirection).
+///
+/// # Argv shape (per ADR-011 §Decision)
+///
+/// `[--print, --agent, <slug>, --output-format, json, --max-budget-usd,
+///   <budget>, [--add-dir, <path>], [--allowedTools, <T1>, <T2>, ...]]`
+///
+/// `--allowedTools` MUST be last because it's variadic — any flag after it
+/// would be consumed as a tool name.
+///
+/// # Security gates (run BEFORE any argv push)
+///
+/// - [`validate_allowed_tools`] — flag-injection guard (R1 audit C-2).
+/// - [`add_dir_for_produces_at`] — path-traversal guard (R1 audit C-1).
+///
+/// Both produce `Err(String)` with a human-readable reason; callers wrap
+/// in `DispatchError::Transport` with the step id for context.
+///
+/// # Errors
+///
+/// Returns `Err(reason: String)` when:
+/// - any `Step.allowed_tools` entry fails `validate_tool_name`;
+/// - `Step.produces_at` has a path-traversal segment (`..` or absolute).
+///
+/// # Example
+///
+/// ```ignore
+/// let argv = build_argv("my-agent", &step, &workspace_root)?;
+/// assert_eq!(argv[0], "--print");
+/// assert_eq!(argv[1], "--agent");
+/// assert_eq!(argv[2], "my-agent");
+/// ```
+pub(super) fn build_argv(
+    slug: &str,
+    step: &Step,
+    workspace_root: &Path,
+) -> Result<Vec<String>, String> {
+    let budget = effective_budget_usd(step);
+    let tools = effective_allowed_tools(step);
+
+    // R1 audit CRITICAL — security-expert C-2: validate every
+    // allowed_tools entry BEFORE argv construction.
+    validate_allowed_tools(&tools)?;
+
+    // R1 audit CRITICAL — security-expert C-1: canonicalise produces_at,
+    // reject `..` and absolute paths to prevent workspace escape via
+    // `--add-dir`.
+    let add_dir = add_dir_for_produces_at(step, workspace_root)?;
+
+    let mut args: Vec<String> = Vec::with_capacity(11 + tools.len());
+    args.push("--print".to_string());
+    args.push("--agent".to_string());
+    args.push(slug.to_string());
+    args.push("--output-format".to_string());
+    args.push("json".to_string());
+    args.push("--max-budget-usd".to_string());
+    // Shared format_budget for argv-shape parity between Plugin/Agent
+    // (pre-fix Plugin emitted "1.00", Agent emitted "1" for default).
+    args.push(format_budget(budget));
+    if let Some(dir) = &add_dir {
+        args.push("--add-dir".to_string());
+        args.push(dir.to_string_lossy().into_owned());
+    }
+    // `--allowedTools` is variadic and MUST be last — any later flag
+    // would be consumed as a tool name.
+    if !tools.is_empty() {
+        args.push("--allowedTools".to_string());
+        for tool in &tools {
+            args.push(tool.clone());
+        }
+    }
+    Ok(args)
+}
+
+/// Full `claude --print` invocation orchestration: argv build + env
+/// allowlist + prompt-via-stdin + spawn + timeout + parse + render.
+///
+/// PROB-050 A-4 closure: extracted from AgentDispatcher::dispatch and
+/// PluginDispatcher::dispatch (identical 9-step recipe per ADR-011
+/// §Decision). Both dispatchers reduce to (a) variant unpack,
+/// (b) compute slug + label, (c) resolve binary, (d) call invoke.
+///
+/// # Arguments
+///
+/// - `label` — dispatcher-prefix string for diagnostics
+///   (`"agent \`foo\`"` or `"plugin \`foo/bar\`"`)
+/// - `slug` — value for `--agent <slug>` argv slot
+/// - `step` — playbook step (provides budget, allowed_tools, produces_at, …)
+/// - `workspace_root` — cwd for the subprocess (relative `produces_at`
+///   paths land here) and root for path-traversal checks
+/// - `binary` — pre-resolved path to the `claude` executable
+/// - `timeout` — pre-resolved per-invocation timeout (already considers
+///   `Step.timeout_seconds` override vs dispatcher default)
+///
+/// # Errors
+///
+/// Returns `DispatchError::Transport` when:
+/// - `build_argv` rejects the step (flag-injection or path-traversal)
+/// - `run_subprocess` cannot spawn the binary (Transport-level)
+///
+/// Returns `Ok(DispatchOutcome { success: false, ... })` when:
+/// - subprocess timed out
+/// - `claude --print` returned a non-success JSON envelope (api error,
+///   budget cap, partial result)
+/// - stdout is not parseable as a [`ClaudePrintResponse`]
+///
+/// Returns `Ok(DispatchOutcome { success: true, ... })` when claude's
+/// JSON envelope decoded with `is_error: false`.
+pub(super) async fn invoke(
+    label: &str,
+    slug: &str,
+    step: &Step,
+    workspace_root: &Path,
+    binary: &Path,
+    timeout: Duration,
+) -> Result<DispatchOutcome, DispatchError> {
+    // 1. Build argv (with security gates: validate_allowed_tools +
+    //    add_dir_for_produces_at).
+    let args = build_argv(slug, step, workspace_root)
+        .map_err(|reason| DispatchError::Transport(format!("{label}: {reason}")))?;
+
+    // 2. Compose env allow-list — base PATH/HOME/USER only. We
+    //    deliberately do NOT forward `ANTHROPIC_API_KEY` etc. — `claude`
+    //    relies on its existing keychain session.
+    let base_env: HashMap<String, String> = std::env::vars().collect();
+    let env = super::helpers::build_env_allowlist(&[], &base_env);
+
+    // 3. Assemble prompt for stdin. produces_at hint is appended by
+    //    `assemble_prompt` itself.
+    let prompt = assemble_prompt(step);
+    let stdin_bytes = prompt.into_bytes();
+
+    // 4. Build subprocess spec.
+    let program_str = binary.to_string_lossy().into_owned();
+    let spec = super::helpers::SubprocessSpec {
+        program: &program_str,
+        args: &args,
+        env: &env,
+        cwd: Some(workspace_root),
+        timeout,
+        stdin_data: Some(&stdin_bytes),
+    };
+
+    // 5. Execute. Helper translates lifecycle into outcome / Transport.
+    let outcome = super::helpers::run_subprocess(spec).await?;
+
+    // 6. Map subprocess outcome → DispatchOutcome.
+    if outcome.timed_out {
+        return Ok(DispatchOutcome {
+            success: false,
+            output_path: None,
+            stderr: Some(format_timeout_msg(label, outcome.duration)),
+        });
+    }
+
+    let stderr_str = if outcome.stderr.is_empty() {
+        String::new()
+    } else {
+        String::from_utf8_lossy(&outcome.stderr).into_owned()
+    };
+
+    match parse_envelope(&outcome.stdout) {
+        Ok(response) => {
+            let success = response.is_success();
+            let stderr = if success {
+                if stderr_str.is_empty() {
+                    None
+                } else {
+                    Some(stderr_str)
+                }
+            } else {
+                let mut combined = response.render_failure_context();
+                if !stderr_str.is_empty() {
+                    combined.push_str(" | stderr=");
+                    combined.push_str(&stderr_str);
+                }
+                Some(combined)
+            };
+            let output_path = if success {
+                step.produces_at.clone()
+            } else {
+                None
+            };
+            Ok(DispatchOutcome {
+                success,
+                output_path,
+                stderr,
+            })
+        }
+        Err(parse_err) => {
+            // Unparseable stdout: JSON envelope is mandatory per ADR-011.
+            // Diagnostic combines exit_code (agent's pattern) + stdout
+            // preview (plugin's pattern) — union of pre-A-4 dispatcher
+            // diagnostics. Prefix uses "failed to decode" wording so the
+            // pre-A-4 plugin test assertion `diag.contains("failed to
+            // decode")` still passes; agent test wasn't string-asserting
+            // this path.
+            let mut diag =
+                format!("{label} failed to decode claude --print JSON envelope: {parse_err}");
+            if let Some(code) = outcome.exit_code {
+                diag.push_str(&format!(" | exit_code={code}"));
+            }
+            if !stderr_str.is_empty() {
+                diag.push_str(" | stderr=");
+                diag.push_str(stderr_str.trim_end());
+            }
+            if !outcome.stdout.is_empty() {
+                let stdout_str = String::from_utf8_lossy(&outcome.stdout);
+                let preview: String = stdout_str.chars().take(MAX_PREVIEW_BYTES).collect();
+                diag.push_str(" | stdout_preview=");
+                diag.push_str(preview.trim_end());
+            }
+            Ok(DispatchOutcome {
+                success: false,
+                output_path: None,
+                stderr: Some(diag),
+            })
+        }
+    }
 }
 
 /// Pattern enforcing argv-injection-safe agent / plugin / target names:
